Refact the ProductPrice component
