+++ CHANGELOG.md
@@ -6,8 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Refact the `ProductPrice` component
 ### Fixed
-- Updatethe proptypes of the components: `BuyButton`, `ProductPrice`, `ProductName`
+- Update the propTypes of the components: `BuyButton`, `ProductPrice`, `ProductName`
 
 ## [2.0.0] - 2018-08-30
 ### Removed
