+++ react/components/ProductPrice/index.js
@@ -1,12 +1,13 @@
 import './global.css'
 
 import PropTypes from 'prop-types'
-import { isEmpty, isNil } from 'ramda'
+import { isNil } from 'ramda'
 import React, { Component } from 'react'
 import ContentLoader from 'react-content-loader'
 import { FormattedMessage, injectIntl } from 'react-intl'
 
 import PricePropTypes from './propTypes'
+import Installments from './Installments'
 
 /**
  * The Price component. Shows the prices information of the Product Summary.
@@ -17,42 +18,6 @@ class Price extends Component {
   }
 
   static propTypes = PricePropTypes
-
-  static Loader = (loaderProps = {}) => (
-    <div className="vtex-price vtex-price-loader">
-      <ContentLoader
-        style={{
-          width: '100%',
-          height: '100%',
-        }}
-        height="100%"
-        width="100%"
-        {...loaderProps}>
-        <rect
-          height="0.75em"
-          width="50%"
-          x="25%"
-          {...loaderProps['vtex-price-list__container--loader']}
-        />
-        <rect {...loaderProps['vtex-price-selling__label--loader']} />
-        <rect
-          height="1em"
-          width="70%"
-          x="15%"
-          y="1.25em"
-          {...loaderProps['vtex-price-selling--loader']}
-        />
-        <rect
-          height="0.75em"
-          width="80%"
-          x="10%"
-          y="2.75em"
-          {...loaderProps['vtex-price-installments--loader']}
-        />
-        <rect {...loaderProps['vtex-price-savings--loader']} />
-      </ContentLoader>
-    </div>
-  )
 
   static defaultProps = {
     showListPrice: true,
@@ -68,77 +33,6 @@ class Price extends Component {
     maximumFractionDigits: 2,
   }
 
-  getInstallmentsNode() {
-    const {
-      installments,
-      showLabels,
-      intl: { formatNumber },
-    } = this.props
-
-    if (!installments || isEmpty(installments)) {
-      return null
-    }
-
-    const noInterestRateInstallments = installments.filter(
-      installment => !installment.InterestRate
-    )
-
-    /*
-     * - The selected installment will be the one with the highest `NumberOfInstallments`;
-     * - If there is no 'interest-free' installments, the normal installments will be analyzed.
-     */
-    const installment = (isEmpty(noInterestRateInstallments)
-      ? installments
-      : noInterestRateInstallments
-    ).reduce(
-      (previous, current) =>
-        previous.NumberOfInstallments > current.NumberOfInstallments
-          ? previous
-          : current
-    )
-
-    const formattedInstallmentPrice = formatNumber(
-      installment.Value,
-      this.currencyOptions
-    )
-
-    const [installmentsElement, installmentPriceElement, timesElement] = [
-      installment.NumberOfInstallments,
-      formattedInstallmentPrice,
-      <span key="times">&times;</span>,
-    ].map((element, index) => (
-      <span className="vtex-price-installments__value" key={index}>
-        {element}
-      </span>
-    ))
-
-    return (
-      <div className="vtex-price-installments__container">
-        <div className="vtex-price-installments dib">
-          {showLabels ? (
-            <FormattedMessage
-              id="pricing.installment-display"
-              values={{
-                installments: installmentsElement,
-                installmentPrice: installmentPriceElement,
-                times: timesElement,
-              }}
-            />
-          ) : (
-              <span>
-                {installmentsElement} {timesElement} {installmentPriceElement}
-              </span>
-            )}
-          {!installment.InterestRate && (
-            <span className="pl1">
-              <FormattedMessage id="pricing.interest-free" />
-            </span>
-          )}
-        </div>
-      </div>
-    )
-  }
-
   render() {
     const {
       sellingPrice,
@@ -147,11 +41,13 @@ class Price extends Component {
       showInstallments,
       showLabels,
       showSavings,
+      installments,
+      styles,
       intl: { formatNumber },
     } = this.props
 
     if ((showListPrice && isNil(listPrice)) || isNil(sellingPrice)) {
-      return <Price.Loader {...this.props.styles} />
+      return <Price.Loader {...styles} />
     }
 
     const differentPrices = showListPrice && sellingPrice !== listPrice
@@ -180,7 +76,13 @@ class Price extends Component {
             {formatNumber(sellingPrice, this.currencyOptions)}
           </div>
         </div>
-        {showInstallments && this.getInstallmentsNode()}
+        {showInstallments &&
+          <Installments
+            installments={installments}
+            showLabels={showLabels}
+            formatNumber={formatNumber}
+            currencyOptions={this.currencyOptions}
+          />}
         {differentPrices &&
           showSavings && (
             <div className="vtex-price-savings__container">
@@ -191,7 +93,7 @@ class Price extends Component {
                     savings: formatNumber(
                       listPrice - sellingPrice,
                       this.currencyOptions
-                    ).replace(" ", ''),
+                    ),
                   }}
                 />
               </div>
@@ -202,9 +104,46 @@ class Price extends Component {
   }
 }
 
-const priceComponent = injectIntl(Price)
+Price.Loader = (loaderProps = {}) => (
+  <div className="vtex-price vtex-price-loader">
+    <ContentLoader
+      style={{
+        width: '100%',
+        height: '100%',
+      }}
+      height="100%"
+      width="100%"
+      {...loaderProps}>
+      <rect
+        height="0.75em"
+        width="50%"
+        x="25%"
+        {...loaderProps['vtex-price-list__container--loader']}
+      />
+      <rect {...loaderProps['vtex-price-selling__label--loader']} />
+      <rect
+        height="1em"
+        width="70%"
+        x="15%"
+        y="1.25em"
+        {...loaderProps['vtex-price-selling--loader']}
+      />
+      <rect
+        height="0.75em"
+        width="80%"
+        x="10%"
+        y="2.75em"
+        {...loaderProps['vtex-price-installments--loader']}
+      />
+      <rect {...loaderProps['vtex-price-savings--loader']} />
+    </ContentLoader>
+  </div>
+)
 
-priceComponent.schema = {
+
+const priceWithIntel = injectIntl(Price)
+
+priceWithIntel.schema = {
   title: 'editor.productPrice.title',
   description: 'editor.productPrice.description',
   type: 'object',
@@ -236,4 +175,4 @@ priceComponent.schema = {
   },
 }
 
-export default priceComponent
+export default priceWithIntel
